requirements.txt
Flask
qrcode
Pillow
dotenv
requests
google-genai
